+++ lib/photox_pro.dart
@@ -3,3 +3,4 @@ library photox_pro;
 export 'src/photox.dart' show PhotoX;
 export 'src/item.dart' show PhotoXItem;
 export 'src/dismissmode.dart' show DismissMode;
+export 'src/cached_image.dart' show CachedImageX;