+++ lib/src/thumbnail.dart
@@ -3,18 +3,23 @@ import 'package:photox_pro/src/cached_image.dart';
 import 'package:photox_pro/src/item.dart';
 
 class PhotoxThumbnail extends StatelessWidget {
+
   const PhotoxThumbnail({
     Key? key,
     required this.item,
     required this.onTap,
     this.thumbnailMargin,
-    this.thumbnailPadding
+    this.thumbnailPadding,
+    this.placeholder,
+    this.errorImage
   }) : super(key: key);
 
   final PhotoXItem item;
   final GestureTapCallback onTap;
   final EdgeInsetsGeometry? thumbnailMargin;
   final EdgeInsetsGeometry? thumbnailPadding;
+  final Widget Function(BuildContext, String)? placeholder;
+  final Widget Function(BuildContext, String, Object)? errorImage;
 
   @override
   Widget build(BuildContext context) {
@@ -30,10 +35,15 @@ class PhotoxThumbnail extends StatelessWidget {
             borderRadius: BorderRadius.circular(6),
             child: item.isAsset
                 ? Image.asset(item.resource, fit: BoxFit.cover)
-                : CachedImageX(url: item.resource),
+                : CachedImageX(
+                  url: item.resource,
+                  placeholder: placeholder,
+                  errorImage: errorImage,
+                ),
           ),
         ),
       ),
     );
   }
+
 }