+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 * Added customizable page indicator for thumbnail gallery
 * Added options to customize fullscreen gallery title TextStyle and leading widget
 * Added gradient behind fullscreen gallery AppBar for better label visibility
+
+## 0.2.5
+
+* Added 'viewportFraction'
+* Added 'thumbnailMargin', 'thumbnailPadding'
+* Added 'placeholder', 'errorImage' (builders)